Runtime: ​bugref:8231 Starting on defining and implementing a new RTIoQueue API to replace RTFileAio mid term. [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@79952 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/ioqueue/ioqueuebase.cpp
@@ -141,8 +141,8 @@ RTDECL(int) RTIoQueueCreate(PRTIOQUEUE phIoQueue, PCRTIOQUEUEPROVVTABLE pProvVTa
     {
         pThis->pVTbl          = pProvVTable;
         pThis->hIoQueueProv   = (RTIOQUEUEPROV)&pThis->abInst[0];
-        pThis->cSqEntries     = cSqEntries;
-        pThis->cCqEntries     = cCqEntries;
+        pThis->cSqEntries     = (uint32_t)cSqEntries;
+        pThis->cCqEntries     = (uint32_t)cCqEntries;
         pThis->cReqsCommitted = 0;
         pThis->cReqsPrepared  = 0;
